controller/requirements.txt
pytest
RPi.GPIO
adafruit-circuitpython-pn532
transitions
black